clean up telemetry
